feat: 타입에 의한 파일 분리
