+++ JuiceMaker/JuiceMaker/Model/FruitStoreError.swift
@@ -1,0 +1,7 @@
+//  JuiceMaker - FruitStoreError.swift
+//  Created by Vetto, 레옹아범 on 2023.1.2
+
+enum FruitStoreError: Error {
+    case lackedInventory
+    case noExistInventory
+}
