+++ JuiceMaker/JuiceMaker/Model/JuiceMaker.swift
@@ -1,37 +1,6 @@
 //  JuiceMaker - JuiceMaker.swift
 //  Created by Vetto, 레옹아범 on 2023.1.2
 
-import Foundation
-
-enum JuiceMenu {
-    case strawberryJuice
-    case bananaJuice
-    case kiwiJuice
-    case pineappleJuice
-    case mangoJuice
-    case strawberryBananaJuice
-    case mangoKiwiJuice
-    
-    var recipe: [Fruit: Int] {
-        switch self {
-        case .strawberryJuice:
-            return [.strawberry: 16]
-        case .bananaJuice:
-            return [.banana: 2]
-        case .kiwiJuice:
-            return [.kiwi: 3]
-        case .pineappleJuice:
-            return [.pineapple: 2]
-        case .mangoJuice:
-            return [.mango: 3]
-        case .strawberryBananaJuice:
-            return [.strawberry: 10, .banana: 1]
-        case .mangoKiwiJuice:
-            return [.mango: 2, .kiwi: 1]
-        }
-    }
-}
-
 // 쥬스 메이커 타입
 struct JuiceMaker {
     let fruitStore = FruitStore()
